Add files via upload
By clicking on the screen, you control the Scuba Diver's up and down movements. As you pass the sharks your score goes up.Keep in mind that the Scuba Diver is falling endlessly so clicking on the screen maneuvers him upwards. If the Scuba Diver touches the borders or the sharks, the game stops.